backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# Pydantic and Validation
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.1.0

# HTTP Client and External APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
xlsxwriter==3.1.9

# Optimization Libraries
ortools==9.8.3296
scipy==1.11.4
scikit-learn==1.3.2

# Geospatial and Routing
geopy==2.4.1
shapely==2.0.2
folium==0.15.0

# Caching and Performance
redis==5.0.1
aioredis==2.0.1
celery==5.3.4

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# File Processing
python-magic==0.4.27
pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Environment and Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0